DAOS-8616 spdk: Performance degradation after repeated format
Currently, doing multiple cycle of "format; create pool/container; run fio"
demonstrates a lot of variability in the results when using NVMe SSD. Adding a
blkdiscard call just before the format seemed to be a temporary fix to the issue.

Apply patch from Jim Harris (SPDK) to align clear operation chunks on a physical
block boundary. UINT32_MAX is not an aligned to IU size when the SSD has 512B
logical format, resulting in the unmap operation being ignored by the SSD.

Signed-off-by: Sydney Vanda sydney.m.vanda@intel.com